SwiftCompilerSources: move the getAccessBaseTest to the SIL module
