+++ SwiftCompilerSources/Sources/Optimizer/Utilities/CMakeLists.txt
@@ -7,7 +7,6 @@
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
 swift_compiler_sources(Optimizer
-  AccessUtilsTest.swift
   AddressUtils.swift
   BorrowUtils.swift
   SpecializationCloner.swift
