+++ SwiftCompilerSources/Sources/Optimizer/Utilities/FunctionTest.swift
@@ -41,7 +41,6 @@ public func registerOptimizerTests() {
 
   // Register each test.
   registerFunctionTests(
-    getAccessBaseTest,
     addressOwnershipLiveRangeTest,
     argumentConventionsTest,
     borrowIntroducersTest,
